pyproject.toml
[project]
name = "abbotttheoctopus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "adafruit-circuitpython-pca9685>=3.4.21",
    "adafruit-circuitpython-rgb-display>=3.14.4",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "setuptools<80",
]
